+++ src/implementations/vanilla/index.html
@@ -15,13 +15,13 @@
   <h1>HIG Example Tests</h1>
   <div>
     <ol>
-      <li><a href="src/basics/button/tests/tests-button.html">button</a></li>
+      <li><a href="src/components/button/tests/tests-button.html">button</a></li>
       <li><a href="src/basics/form-elements/checkbox/tests/tests-checkbox.html">checkbox</a></li>
       <li><a href="src/basics/colors/tests/tests-colors.html">colors</a></li>
       <li><a href="src/basics/form-elements/dropdown/tests/tests-dropdown.html">dropdown</a></li>
       <li><a href="src/basics/flyout/tests/test-flyout.html">flyout</a></li>
       <li><a href="src/components/global-nav/tests/tests-global-nav-bim.html">global nav bim</a></li>
-      <li><a href="src/basics/icon-button/tests/tests-icon-button.html">icon button</a></li>
+      <li><a href="src/components/icon-button/tests/tests-icon-button.html">icon button</a></li>
       <li><a href="src/components/modal/tests/tests-modal.html">modal</a></li>
       <li><a href="src/basics/form-elements/password-field/tests/tests-password-field.html">password field</a></li>
       <li><a href="src/basics/form-elements/radio-button/tests/tests-radio-button.html">radio button</a></li>
