+++ src/implementations/vanilla/src/components/icon-button/icon-button.html
